requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
requests
python-dotenv
apscheduler
opencv-python-headless
numpy
psycopg2-binary
pytest
oss2
